Collecting message related stuff into message folder
