+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/message/attachment/Utils.kt
@@ -1,0 +1,7 @@
+package io.getstream.chat.android.offline.message.attachment
+
+import java.util.UUID
+
+internal fun generateUploadId(): String {
+    return "upload_id_${UUID.randomUUID()}"
+}
