+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/offline/channel/controller/WhenHandleEvent.kt
@@ -13,7 +13,7 @@ import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.offline.ChatDomainImpl
 import io.getstream.chat.android.offline.channel.ChannelController
-import io.getstream.chat.android.offline.channel.MessageHelper
+import io.getstream.chat.android.offline.message.attachment.AttachmentUrlValidator
 import io.getstream.chat.android.offline.randomChannel
 import io.getstream.chat.android.offline.randomChannelDeletedEvent
 import io.getstream.chat.android.offline.randomChannelUpdatedEvent
@@ -50,13 +50,13 @@ internal class WhenHandleEvent {
         on(it.currentUser) doReturn currentUser
         on(it.getChannelConfig(any())) doReturn Config(isConnectEvents = true, isMutes = true)
     }
-    private val messageHelper: MessageHelper = mock()
+    private val attachmentUrlValidator: AttachmentUrlValidator = mock()
 
     private lateinit var channelController: ChannelController
 
     @BeforeEach
     fun setUp() {
-        whenever(messageHelper.updateValidAttachmentsUrl(any(), any())) doAnswer { invocation ->
+        whenever(attachmentUrlValidator.updateValidAttachmentsUrl(any(), any())) doAnswer { invocation ->
             invocation.arguments[0] as List<Message>
         }
 
@@ -65,7 +65,7 @@ internal class WhenHandleEvent {
             channelId = channelId,
             client = chatClient,
             domainImpl = chatDomain,
-            messageHelper = messageHelper
+            messageHelper = attachmentUrlValidator
         )
     }
 
@@ -81,7 +81,7 @@ internal class WhenHandleEvent {
             showInChannel = true
         )
 
-        whenever(messageHelper.updateValidAttachmentsUrl(any(), any())) doReturn listOf(newMessage)
+        whenever(attachmentUrlValidator.updateValidAttachmentsUrl(any(), any())) doReturn listOf(newMessage)
 
         val userStartWatchingEvent = randomNewMessageEvent(user = user, createdAt = newDate, message = newMessage)
 
@@ -101,7 +101,7 @@ internal class WhenHandleEvent {
             showInChannel = true
         )
         val newMessageEvent = randomNewMessageEvent(user = user, message = message)
-        whenever(messageHelper.updateValidAttachmentsUrl(any(), any())) doReturn listOf(message)
+        whenever(attachmentUrlValidator.updateValidAttachmentsUrl(any(), any())) doReturn listOf(message)
 
         channelController.handleEvent(newMessageEvent)
 
@@ -257,7 +257,7 @@ internal class WhenHandleEvent {
             silent = false,
         )
         val reactionEvent = randomReactionNewEvent(user = randomUser(), message = message)
-        whenever(messageHelper.updateValidAttachmentsUrl(any(), any())) doReturn listOf(message)
+        whenever(attachmentUrlValidator.updateValidAttachmentsUrl(any(), any())) doReturn listOf(message)
 
         channelController.handleEvent(reactionEvent)
