+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/offline/channel/controller/attachment/AttachmentUploaderTests.kt
@@ -14,7 +14,7 @@ import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.errors.ChatError
 import io.getstream.chat.android.client.models.Attachment
 import io.getstream.chat.android.client.utils.Result
-import io.getstream.chat.android.offline.channel.attachment.AttachmentUploader
+import io.getstream.chat.android.offline.message.attachment.AttachmentUploader
 import io.getstream.chat.android.offline.randomAttachmentsWithFile
 import io.getstream.chat.android.offline.randomMessage
 import io.getstream.chat.android.test.TestCall
